+++ projectroles/forms.py
@@ -117,8 +117,32 @@ class SODARModelForm(SODARFormMixin, forms.ModelForm):
 
 
 class SODARPagedownWidget(PagedownWidget):
+    """Pagedown widget for markdown fields"""
+
     class Media:
         css = {'all': ['projectroles/css/pagedown.css']}
+
+
+class MultipleFileInput(forms.ClearableFileInput):
+    """Multiple file input widget working with Django 3.2.19+"""
+
+    allow_multiple_selected = True
+
+
+class MultipleFileField(forms.FileField):
+    """Multiple file input field working with Django 3.2.19+"""
+
+    def __init__(self, *args, **kwargs):
+        kwargs.setdefault('widget', MultipleFileInput())
+        super().__init__(*args, **kwargs)
+
+    def clean(self, data, initial=None):
+        single_file_clean = super().clean
+        if isinstance(data, (list, tuple)):
+            result = [single_file_clean(d, initial) for d in data]
+        else:
+            result = single_file_clean(data, initial)
+        return result
 
 
 # User autocompletion ----------------------------------------------------------