+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 - **Projectroles**
     - ``TestSiteAppPermissionBase`` base test class (#1236)
     - ``full_title`` arg in ``Project.get_log_title()`` (#1238)
+    - ``MultipleFileInput`` and ``MultipleFileField`` form helpers (#1226)
 
 Changed
 -------